sample_market_data_project/data-generator/requirements.txt
kafka-python==2.0.2
requests==2.27.1
schedule==1.1.0
yfinance